refactor: upd naming
